requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
jinja2==3.1.4
aiofiles==24.1.0
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
itsdangerous==2.2.0
aiosqlite==0.20.0
httpx==0.28.1
python-dotenv==1.0.1
slowapi==0.1.9
starlette==0.41.3
